use int64_t instead of int

+++ main.cpp
@@ -4,7 +4,7 @@
 
 
 
-// int cost(Utuple<3,char> t) {
+// int64_t cost(Utuple<3,char> t) {
 //     // if (t[0] == 'a' && t[1] == 'b' && t[2] == 'c')
 //     //     return -5;
 //     if (t[0] == 'b' && t[1] == 'c' && t[2] == 'd')
@@ -16,7 +16,7 @@
 //     return 0;
 // }
 
-// int cost(Utuple<3,char> t) {
+// int64_t cost(Utuple<3,char> t) {
 //     // example: 3.1 and 3.11 are sufficient
 //     if (t[0] == 'a' && t[1] == 'b' && t[2] == 'c') return -1;
 //     if (t[0] == 'a' && t[1] == 'c' && t[2] == 'd') return -15;
@@ -30,7 +30,7 @@
 // }
 
 
-int cost(Utuple<3,char> t) {
+int64_t cost(Utuple<3,char> t) {
     // // pyramid example below (3.1 + 3.11 are not sufficient) (3.4 is sufficient)
     // if (t[0] == 'a' && t[1] == 'b' && t[2] == 'e') return -75;
     // pyramid example below (3.1 + 3.11 + 3.4 are not sufficient) (3.6 is sufficient for 10) (everything is insufficient for 70, 3.3 separates the pyramide base for 100)
